Add files via upload
#INTRODUCTION
The main objective of the project is to develop online Banking system for banks. In present 
system all banking work is done manually. User have to visit bank to Withdrawal or Deposit 
amount. In present bank system it is also difficult to find account information of account 
holder. In this bank management system we will automate all the banking process. In our 
bank management system user can check his balance online and he can also transfer money 
to other account online. In this Software you can keep record for daily Banking transactions. 
The main purpose of developing bank management system is to design an application, which 
could store bank data and provide an interface for retrieving customer related details with 
100% accuracy.
This bank management system also allow user to add new customer account, delete account 
and user can also modify existing user account information. Using this system user can also 
search any individual account in few seconds. Using our bank management system user can 
also check any translation in any account. Our system also provide security check to reduce 
fraud. The system will check the user’s existence in the database and provide the set of 
services with respect to the role of the user.
#PROPOSED SYSTEM
Proposed system is an automated Bank Management System. Through our software user can 
add account, deposit, withdraw, display account list in quick time. Our proposed system has 
the following advantages.
➢ User friendly interface
➢ Fast access to database
➢ Less error
➢ More Storage Capacity
➢ Search facility
➢ Look and Feel Environment
➢ Quick transaction
All the manual difficulties in managing a Bank have been rectified by implementing 
computerization.
#FEASIBILITY ANALYSIS
Whatever we think need not be feasible .It is wise to think about the feasibility of any problem 
we undertake. Feasibility is the study of impact, which happens in the organization by the 
development of a system. The impact can be either positive or negative. When the positives 
nominate the negatives, then the system is considered feasible. Here the feasibility study can 
be performed in two ways such as technical feasibility and Economical Feasibility. 
1.Technical Feasibility:
 We can strongly says that it is technically feasible, since there will not be much difficulty 
in getting required resources for the development and maintaining the system as well. All the 
resources needed for the development of the software as well as the maintenance of the 
same is available in the organization here we are utilizing the resources which are available 
already.
2.Economical Feasibility 
 Development of this application is highly economically feasible .The organization needed 
not spend much m one for the development of the system already available. The only thing 
is to be done is making an environment for the development with an effective supervision. If 
we are doing so , we can attain the maximum usability of the corresponding resources .Even 
after the development , the organization will not be in a condition to invest more in the 
organization .Therefore , the system is economically feasible.